fix(skillchain): record SC damage as int32 (int16 capped hits at 32,767)
recordSkillchain took int16_t dmg, truncating skillchain damage at 32,767. Widen to int32_t so SC records up to the 131,071 packet cap (pairs with the magic/WS uncap). Full SC HP-uncap would also need TakeSkillchainDamage in battleutils.cpp.

+++ src/map/action/action.cpp
@@ -22,7 +22,7 @@
 #include "action.h"
 #include "utils/battleutils.h"
 
-void action_result_t::recordSkillchain(const ActionProcSkillChain effect, const int16_t dmg)
+void action_result_t::recordSkillchain(const ActionProcSkillChain effect, const int32_t dmg)
 {
     if (dmg < 0)
     {
